// src/app/types/types.ts
export interface IItemStop {
  DateArrive: string;
  TimeArrive: string;
  City: string;
  Name: string;
  Id?: null;
  Latitude: null | string;
  Longitude: null | string;
}
export interface IItemBusOptions {
  Id?: number;
  Name: string;
  IsEnabled: boolean;
  Description?: null | string;
  Link?: null | string;
}
export interface IItemFullBusPlaces {
  Seat: number;
  IsFree: boolean;
  Row: number;
  Col: number;
  Floor: number;
}
export interface IItemPrice {
  Ptar: number;
  Etar: number;
  Dtar: number;
  Currency?: number;
  CurrencyName?: string;
}
export interface IItemBaggagePrices {
  Currency?: number;
  CurrencyName?: string;
  Value?: number;
}
export interface IItemRoutesProps {
  GuidIdentityNumber?: null;
  City1?: {
    Id: number;
    NameEng?: string;
    NameRus: string;
  };
  CitySourceId: number;
  CityTargetId: number;
  CitySourceUniqueId?: null;
  CityTargetUniqueId?: null;
  CityCodeSource?: null;
  CityCodeTarget?: null;
  StopNameSource?: null;
  StopNameTarget?: null;
  City2: {
    Id: number;
    NameEng?: string;
    NameRus: string;
  };
  Path: {
    Id: number;
    Name?: string;
    CarrierId: number;
  };
  Date: string;
  Count: number;
  FreeSeats?: null;
  RequestGetFreeSeats?: null;
  DiscountListCode?: null;
  TripCode?: null;
  TimeOfArrival?: null;
  TimeOfDeparture?: null;
  SystemId?: null;
  CarrierName?: null;
  CarrierId?: string;
  Number?: null;
  RealCarrierId?: number;
  Tariffs?: null;
  Byn?: number;
  IsChangeRoute?: boolean;
  ChangeRouteDetails?: null;
}

export interface IItemRoutes {
  AddInfo: null | number | string;
  AdditionalPlace?: null | number;
  AllStops: IItemStop[];
  AnnylationRules?: string;
  ArriveDateTime?: string;
  ArriveName: string;
  BaggageDescription?: string;
  BaggagePrices?: IItemBaggagePrices[];
  BusId?: null | number;
  BusOptions?: IItemBusOptions[];
  BusPlaces?: null | number;
  Carrier: number;
  CarrierLogo?: null;
  CarrierName: string;
  City1: string;
  City2: string;
  CountFreePromos: number;
  DateArrive: string;
  DateArrive1?: string;
  DateDepart: string;
  DateDepart1?: string;
  Delete?: string;
  DepartDateTime?: string;
  DepartName: string;
  FullBusPlaces: IItemFullBusPlaces[];
  FullPrice?: null | string | number;
  FreePlace?: number;
  FreePlaces?: null | number;
  HasExpressAttr?: boolean;
  Hour: string;
  Id: string;
  IsCache: boolean;
  IsInternationalAbroad?: boolean;
  IsPartnerRoute: boolean;
  IsSpecialRules?: boolean;
  Link?: null;
  MaxDiscountPercent?: null;
  Minuts: string;
  Path?: string;
  Platforma?: string;
  Price: IItemPrice[];
  RealCarrierId?: number;
  RealCarrierName: string;
  RealIsPersonalCarrier?: boolean;
  Route: string;
  RouteNumber?: null;
  Routes: IItemRoutesProps[];
  SessionRouteId?: null;
  SortPriority?: number;
  TicketRules?: null | string;
  TimeArrive: string;
  TimeDepart: string;
  UsedByArtmark?: null | string;
}